GraphTeeth/requirements.txt
easydict==1.13
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==11.1.0
pycocotools==2.0.7
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.3.2
terminaltables==3.1.10
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.65.2
